Rename numpy_test to scipy_test and add scipy test suite.
Because scipy rely on NumPy, so add scipy test to numpy_test.py. And
rename it to scipy_test.py. numpy_full_test.py will check numpy test
suite. The scipy_test.py only for check scipy test suite. But scipy test
will take a long time. So disabled it by default.
